requirements.txt
packit == 0.23
apache-airflow
psycopg2-binary
werkzeug
flask-bcrypt